+++ R/utils.R
@@ -118,28 +118,35 @@ is_app <- function(path) {
   )
 }
 
-# Given a path, return a path that can be passed to ShinyDriver$new()
-# * If it is a path to an Rmd file including filename (like foo/doc.Rmd), return path unchanged.
-# * If it is a dir containing app.R, server.R, return path unchanged.
-# * If it is a dir containing index.Rmd, return the path with index.Rmd at the end.
-# * Otherwise, throw error.
-app_path <- function(path) {
-  if (grepl("\\.Rmd", path, ignore.case = TRUE)) {
-    return(path)
+app_path <- function(path, arg = "path") {
+  if (!file.exists(path)) {
+    stop(paste0("'", path, "' doesn't exist"), call. = FALSE)
   }
-  if (dir_exists(path)) {
-    if (any(c("app.r", "server.r") %in% tolower(dir(path)))) {
-      return(path)
+
+  if (is_app(path)) {
+    app <- path
+    dir <- path
+  } else if (is_rmd(path)) {
+    # Fallback for old behaviour
+    if (length(dir(dirname(path), pattern = "\\.[Rr]md$")) > 1) {
+      abort("For testing, only one .Rmd file is allowed per directory.")
     }
-    if ("index.Rmd" %in% dir(path)) {
-      return(file.path(path, "index.Rmd"))
+    app <- path
+    dir <- dirname(path)
+  } else {
+    rmds <- dir(path, pattern = "\\.Rmd$", full.names = TRUE)
+    if (length(rmds) != 1) {
+      abort(paste0(
+        "`", arg, "` doesn't contain 'app.R', 'server.R', or exactly one '.Rmd'"
+      ))
+    } else {
+      app <- rmds
+      dir <- dirname(app)
     }
   }
 
-  abort(paste0(path, " must be a directory containing app.R, server.R, or index.Rmd; or path to a .Rmd file (including the filename)."))
+  list(app = app, dir = dir)
 }
-
-
 
 raw_to_utf8 <- function(data) {
   res <- rawToChar(data)