+++ tests/testthat/test-shiny-driver.R
@@ -38,6 +38,14 @@ test_that("window size", {
   )
 })
 
+test_that("can change pass render_args to rmarkdown::run()", {
+  doc <- ShinyDriver$new(
+    test_path("apps/render-args/doc.Rmd"),
+    renderArgs = list(params = list(name = "Mary"))
+  )
+  expect_equal(doc$getValue("test"), "Mary")
+})
+
 test_that("useful error message if app terminated", {
   skip_on_os("windows") # errors with "Empty reply from server"