+++ NEWS.md
@@ -1,6 +1,10 @@
 shinytest 1.4.0.9002
 ===============
 
+* `ShinyDriver$intialize()` gains a `renderArgs` method which is passed along
+  to `rmarkdown::run()` for interactive Rmds (#249); this makes it possible
+  to set parameters for parameterised reports.
+
 * `ShinyDriver$getAllValues()`, `ShinyDriver$snapshot()`, and
   `ShinyDriver$snapshotDownload()` give clear errors messages if the Shiny
   app is no longer running (e.g. because you've trigged a `stopApp()`) (#192).